Desktop/multimodal-healthcare-assistant/requirements.txt
opencv-python
mediapipe 
numpy 
sounddevice 
openai-whisper 
transformers 
torch 
feat